+++ source/jquery.flot.navigate.js
@@ -111,8 +111,99 @@ import { browser } from './jquery.flot.browser.js';
 import { uiConstants } from './jquery.flot.uiConstants.js';
 import { bind, unbind, trigger, css } from './helpers.js';
 
+/** @typedef {'x' | 'y'} AxisDirection */
+/** @typedef {'manual' | 'smart' | 'smartLock'} PanMode */
+/** @typedef {{ x: number, y: number }} PixelDelta */
+/** @typedef {{ left: number, top: number }} PlotPoint */
+/** @typedef {{ left: number, top: number }} PlotPageOffset */
+/** @typedef {{ left: number, right: number, top: number, bottom: number }} PlotOffset */
+/** @typedef {[number | undefined, number | undefined]} AxisRange */
+
+/**
+ * @typedef {Object} NavigationAxisOptions
+ * @property {boolean} axisZoom
+ * @property {boolean} plotZoom
+ * @property {boolean} axisPan
+ * @property {boolean} plotPan
+ * @property {AxisRange} panRange
+ * @property {AxisRange} zoomRange
+ * @property {{ below?: number, above?: number }} offset
+ */
+
+/**
+ * @typedef {Object} NavigationAxis
+ * @property {AxisDirection} direction
+ * @property {number} min
+ * @property {number} max
+ * @property {{ left: number, top: number, width: number, height: number }} [box]
+ * @property {NavigationAxisOptions} options
+ * @property {(value: number | undefined) => number} p2c
+ * @property {(value: number) => number} c2p
+ */
+
+/** @typedef {NavigationAxis[] | Record<string, NavigationAxis>} AxisCollection */
+
+/**
+ * @typedef {Object} NavigationOptions
+ * @property {{ interactive: boolean, active: boolean, amount: number }} zoom
+ * @property {{ interactive: boolean, active: boolean, cursor: string, frameRate: number | null, mode: PanMode }} pan
+ * @property {{ interactive: boolean }} recenter
+ * @property {Omit<NavigationAxisOptions, 'offset'>} xaxis
+ * @property {Omit<NavigationAxisOptions, 'offset'>} yaxis
+ */
+
+/**
+ * @typedef {Object} NavigationAxisState
+ * @property {{ below: number, above: number }} navigationOffset
+ * @property {number} axisMin
+ * @property {number} axisMax
+ * @property {boolean} diagMode
+ */
+
+/**
+ * @typedef {{ startPageX: number, startPageY: number, diagMode?: boolean }
+ *   & Record<string, NavigationAxisState | number | boolean>} NavigationState
+ */
+
+/** @typedef {{ amount?: number | null, center?: PlotPoint, axes?: AxisCollection, preventEvent?: boolean }} ZoomArguments */
+/** @typedef {{ left: number, top: number, axes?: AxisCollection | null, preventEvent?: boolean }} PanArguments */
+/** @typedef {{ axes?: AxisCollection | null }} RecenterArguments */
+/** @typedef {{ start: PixelDelta, end: PixelDelta | false }} PanHint */
+/** @typedef {((delta: PixelDelta, initialState: NavigationState, panAxes?: AxisCollection | null, preventEvent?: boolean, smartLock?: boolean) => void) & { end: () => void }} SmartPan */
+
+/**
+ * @typedef {Object} NavigationPlot
+ * @property {() => NavigationOptions} getOptions
+ * @property {() => HTMLElement} getPlaceholder
+ * @property {() => PlotPageOffset} offset
+ * @property {() => PlotOffset} getPlotOffset
+ * @property {() => number} width
+ * @property {() => number} height
+ * @property {() => Record<string, NavigationAxis>} getAxes
+ * @property {() => NavigationAxis[]} getXAxes
+ * @property {() => NavigationAxis[]} getYAxes
+ * @property {(redraw?: boolean) => void} setupGrid
+ * @property {() => void} draw
+ * @property {() => void} triggerRedrawOverlay
+ * @property {() => void} activate
+ * @property {(args?: ZoomArguments) => void} zoom
+ * @property {(args?: ZoomArguments) => void} zoomOut
+ * @property {(args: PanArguments) => void} pan
+ * @property {(args: RecenterArguments) => void} recenter
+ * @property {SmartPan} smartPan
+ * @property {(startPageX?: number, startPageY?: number) => NavigationState} navigationState
+ * @property {(touchPointX: number, touchPointY: number) => NavigationAxis[]} getTouchedAxis
+ * @property {{
+ *   processOptions: Array<(plot: NavigationPlot, options: NavigationOptions) => void>,
+ *   bindEvents: Array<(plot: NavigationPlot, eventHolder: HTMLElement) => void>,
+ *   shutdown: Array<(plot: NavigationPlot, eventHolder: HTMLElement) => void>,
+ *   drawOverlay: Array<(plot: NavigationPlot, ctx: CanvasRenderingContext2D) => void>
+ * }} hooks
+ */
+
     'use strict';
 
+    /** @type {NavigationOptions} */
     var options = {
         zoom: {
             interactive: false,
@@ -150,17 +241,23 @@ import { bind, unbind, trigger, css } from './helpers.js';
     var SNAPPING_CONSTANT = uiConstants.SNAPPING_CONSTANT;
     var PANHINT_LENGTH_CONSTANT = uiConstants.PANHINT_LENGTH_CONSTANT;
 
+    /** @param {NavigationPlot} plot */
     function init(plot) {
         plot.hooks.processOptions.push(initNevigation);
     }
 
+    /** @param {NavigationPlot} plot @param {NavigationOptions} options */
     function initNevigation(plot, options) {
+        /** @type {NavigationAxis[] | null | undefined} */
         var panAxes = null;
+        /** @type {HTMLElement} */
+        var navigationEventHolder;
         var canDrag = false;
         var useManualPan = options.pan.mode === 'manual',
             smartPanLock = options.pan.mode === 'smartLock',
             useSmartPan = smartPanLock || options.pan.mode === 'smart';
 
+        /** @param {MouseEvent} e @param {boolean} zoomOut @param {number | null} amount */
         function onZoomClick(e, zoomOut, amount) {
             var page = browser.getPageXY(e);
 
@@ -202,13 +299,20 @@ import { bind, unbind, trigger, css } from './helpers.js';
         }
 
         var prevCursor = 'default',
+            /** @type {PanHint | null} */
             panHint = null,
+            /** @type {ReturnType<typeof setTimeout> | null} */
             panTimeout = null,
+            /** @type {NavigationState} */
             plotState,
             prevDragPosition = { x: 0, y: 0 },
             isPanAction = false;
 
+        /** @param {Event} e */
         function onMouseWheel(e) {
+            if (!(e instanceof WheelEvent)) {
+                return undefined;
+            }
             var delta = -e.deltaY;
             var maxAbsoluteDeltaOnMac = 1,
                 isMacScroll = Math.abs(e.deltaY) <= maxAbsoluteDeltaOnMac,
@@ -230,7 +334,8 @@ import { bind, unbind, trigger, css } from './helpers.js';
 
         plot.navigationState = function(startPageX, startPageY) {
             var axes = this.getAxes();
-            var result = {};
+            /** @type {NavigationState} */
+            var result = { startPageX: startPageX || 0, startPageY: startPageY || 0 };
             Object.keys(axes).forEach(function(axisName) {
                 var axis = axes[axisName];
                 result[axisName] = {
@@ -242,11 +347,10 @@ import { bind, unbind, trigger, css } from './helpers.js';
                 }
             });
 
-            result.startPageX = startPageX || 0;
-            result.startPageY = startPageY || 0;
             return result;
         }
 
+        /** @param {MouseEvent} e */
         function onDragStart(e) {
 
             isPanAction = true;
@@ -285,6 +389,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         }
 
+        /** @param {MouseEvent} e */
         function onDrag(e) {
             if (!isPanAction) {
                 return;
@@ -335,6 +440,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }, 1 / frameRate * 1000);
         }
 
+        /** @param {MouseEvent} e */
         function onDragEnd(e) {
             if (!isPanAction) {
                 return;
@@ -367,7 +473,11 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         }
 
+        /** @param {Event} e */
         function onDblClick(e) {
+            if (!(e instanceof MouseEvent)) {
+                return;
+            }
             plot.activate();
             var o = plot.getOptions()
 
@@ -386,7 +496,11 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         }
 
+        /** @param {Event} e */
         function onClick(e) {
+            if (!(e instanceof MouseEvent)) {
+                return undefined;
+            }
             plot.activate();
 
             if (isPanAction) {
@@ -405,18 +519,24 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         }
 
+        /** @param {Event} e */
         function onPointerDown(e) {
+            if (!(e instanceof PointerEvent)) {
+                return;
+            }
             if (e.button !== 0) {
                 return;
             }
-            var el = e.currentTarget;
+            var el = navigationEventHolder;
             canDrag = true;
             onDragStart(e);
 
+            /** @param {PointerEvent} e */
             function onPointerMove(e) {
                 onDrag(e);
             }
 
+            /** @param {PointerEvent} e */
             function onPointerUp(e) {
                 onDragEnd(e);
                 canDrag = false;
@@ -432,7 +552,9 @@ import { bind, unbind, trigger, css } from './helpers.js';
             el.addEventListener("pointercancel", onPointerUp);
         }
 
+        /** @param {NavigationPlot} plot @param {HTMLElement} eventHolder */
         function bindEvents(plot, eventHolder) {
+            navigationEventHolder = eventHolder;
             var o = plot.getOptions();
             if (o.zoom.interactive) {
                 bind(eventHolder, "wheel", onMouseWheel);
@@ -446,6 +568,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             bind(eventHolder, "click", onClick);
         }
 
+        /** @param {ZoomArguments} [args] */
         plot.zoomOut = function(args) {
             if (!args) {
                 args = {};
@@ -459,6 +582,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             plot.zoom(args);
         };
 
+        /** @param {ZoomArguments} [args] */
         plot.zoom = function(args) {
             if (!args) {
                 args = {};
@@ -490,12 +614,8 @@ import { bind, unbind, trigger, css } from './helpers.js';
                     }
                 };
 
-            for (var key in axes) {
-                if (!axes.hasOwnProperty(key)) {
-                    continue;
-                }
-
-                var axis = axes[key],
+            Object.values(axes).forEach(function(axis) {
+                var
                     opts = axis.options,
                     min = minmax[axis.direction].min,
                     max = minmax[axis.direction].max,
@@ -503,7 +623,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
 
                 //skip axis without axisZoom when zooming only on certain axis or axis without plotZoom for zoom on entire plot
                 if ((!opts.axisZoom && args.axes) || (!args.axes && !opts.plotZoom)) {
-                    continue;
+                    return;
                 }
 
                 min = saturated.saturate(axis.c2p(min));
@@ -519,18 +639,18 @@ import { bind, unbind, trigger, css } from './helpers.js';
                 if (opts.zoomRange) {
                     // zoomed in too far
                     if (max - min < opts.zoomRange[0]) {
-                        continue;
+                        return;
                     }
                     // zoomed out to far
                     if (max - min > opts.zoomRange[1]) {
-                        continue;
+                        return;
                     }
                 }
 
                 var offsetBelow = saturated.saturate(navigationOffset.below - (axis.min - min));
                 var offsetAbove = saturated.saturate(navigationOffset.above - (axis.max - max));
                 opts.offset = { below: offsetBelow, above: offsetAbove };
-            };
+            });
 
             plot.setupGrid(true);
             plot.draw();
@@ -540,6 +660,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         };
 
+        /** @param {PanArguments} args */
         plot.pan = function(args) {
             var delta = {
                 x: +args.left,
@@ -554,8 +675,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
 
             var panAxesOrAll = args.axes || plot.getAxes();
-            Object.keys(panAxesOrAll).forEach(function(key) {
-                var axis = panAxesOrAll[key];
+            Object.values(panAxesOrAll).forEach(function(axis) {
                 var opts = axis.options,
                     d = delta[axis.direction];
 
@@ -613,10 +733,10 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         };
 
+        /** @param {RecenterArguments} args */
         plot.recenter = function(args) {
             var recenterAxes = args.axes || plot.getAxes();
-            Object.keys(recenterAxes).forEach(function(key) {
-                var axis = recenterAxes[key];
+            Object.values(recenterAxes).forEach(function(axis) {
                 if (args.axes) {
                     if (axis.direction === 'x') {
                         axis.options.offset = { below: 0 };
@@ -631,6 +751,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             plot.draw();
         };
 
+        /** @param {PixelDelta} delta */
         var shouldSnap = function(delta) {
             return (Math.abs(delta.y) < SNAPPING_CONSTANT && Math.abs(delta.x) >= SNAPPING_CONSTANT) ||
                 (Math.abs(delta.x) < SNAPPING_CONSTANT && Math.abs(delta.y) >= SNAPPING_CONSTANT);
@@ -638,6 +759,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
 
         // adjust delta so the pan action is constrained on the vertical or horizontal direction
         // it the movements in the other direction are small
+        /** @param {PixelDelta} delta */
         var adjustDeltaToSnap = function(delta) {
             if (Math.abs(delta.x) < SNAPPING_CONSTANT && Math.abs(delta.y) >= SNAPPING_CONSTANT) {
                 return {x: 0, y: delta.y};
@@ -650,7 +772,9 @@ import { bind, unbind, trigger, css } from './helpers.js';
             return delta;
         }
 
+        /** @type {AxisDirection | null} */
         var lockedDirection = null;
+        /** @param {PixelDelta} delta */
         var lockDeltaDirection = function(delta) {
             if (!lockedDirection && Math.max(Math.abs(delta.x), Math.abs(delta.y)) >= SNAPPING_CONSTANT) {
                 lockedDirection = Math.abs(delta.x) < Math.abs(delta.y) ? 'y' : 'x';
@@ -666,6 +790,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         }
 
+        /** @param {PixelDelta} delta */
         var isDiagonalMode = function(delta) {
             if (Math.abs(delta.x) > 0 && Math.abs(delta.y) > 0) {
                 return true;
@@ -673,20 +798,32 @@ import { bind, unbind, trigger, css } from './helpers.js';
             return false;
         }
 
+        /** @param {AxisCollection} axes @param {NavigationState} initialState @param {PixelDelta} delta */
         var restoreAxisOffset = function(axes, initialState, delta) {
-            var axis;
-            Object.keys(axes).forEach(function(axisName) {
-                axis = axes[axisName];
+            Object.entries(axes).forEach(function(entry) {
+                var axisName = entry[0], axis = entry[1];
                 if (delta[axis.direction] === 0) {
-                    axis.options.offset.below = initialState[axisName].navigationOffset.below;
-                    axis.options.offset.above = initialState[axisName].navigationOffset.above;
+                    var axisState = initialState[axisName];
+                    if (typeof axisState !== 'object') {
+                        return;
+                    }
+                    axis.options.offset.below = axisState.navigationOffset.below;
+                    axis.options.offset.above = axisState.navigationOffset.above;
                 }
             });
         }
 
         var prevDelta = { x: 0, y: 0 };
-        plot.smartPan = function(delta, initialState, panAxes, preventEvent, smartLock) {
+        /**
+         * @param {PixelDelta} delta
+         * @param {NavigationState} initialState
+         * @param {AxisCollection | null} [panAxes]
+         * @param {boolean} [preventEvent]
+         * @param {boolean} [smartLock]
+         */
+        function smartPan(delta, initialState, panAxes, preventEvent, smartLock) {
             var snap = smartLock ? true : shouldSnap(delta),
+                /** @type {AxisCollection} */
                 axes = plot.getAxes(),
                 opts;
             delta = smartLock ? lockDeltaDirection(delta) : adjustDeltaToSnap(delta);
@@ -732,9 +869,8 @@ import { bind, unbind, trigger, css } from './helpers.js';
                 axes = panAxes;
             }
 
-            var axis, axisMin, axisMax, p, d;
-            Object.keys(axes).forEach(function(axisName) {
-                axis = axes[axisName];
+            var axisMin, axisMax, p, d;
+            Object.values(axes).forEach(function(axis) {
                 axisMin = axis.min;
                 axisMax = axis.max;
                 opts = axis.options;
@@ -790,15 +926,17 @@ import { bind, unbind, trigger, css } from './helpers.js';
             if (!preventEvent) {
                 trigger(plot.getPlaceholder(), "plotpan", [plot, delta, panAxes, initialState]);
             }
-        };
+        }
 
-        plot.smartPan.end = function() {
+        smartPan.end = function() {
             panHint = null;
             lockedDirection = null;
             prevDelta = { x: 0, y: 0 };
             plot.triggerRedrawOverlay();
         }
+        plot.smartPan = smartPan;
 
+        /** @param {NavigationPlot} plot @param {HTMLElement} eventHolder */
         function shutdown(plot, eventHolder) {
             unbind(eventHolder, "wheel", onMouseWheel);
             unbind(eventHolder, "pointerdown", onPointerDown);
@@ -810,27 +948,14 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         }
 
+        /** @param {NavigationPlot} plot @param {CanvasRenderingContext2D} ctx */
         function drawOverlay(plot, ctx) {
             if (panHint) {
                 ctx.strokeStyle = 'rgba(96, 160, 208, 0.7)';
                 ctx.lineWidth = 2;
                 ctx.lineJoin = "round";
                 var startx = Math.round(panHint.start.x),
-                    starty = Math.round(panHint.start.y),
-                    endx = 0, endy = 0;
-
-                if (panAxes) {
-                    if (panAxes[0].direction === 'x') {
-                        endy = Math.round(panHint.start.y);
-                        endx = Math.round(panHint.end.x);
-                    } else if (panAxes[0].direction === 'y') {
-                        endx = Math.round(panHint.start.x);
-                        endy = Math.round(panHint.end.y);
-                    }
-                } else {
-                    endx = Math.round(panHint.end.x);
-                    endy = Math.round(panHint.end.y);
-                }
+                    starty = Math.round(panHint.start.y);
 
                 ctx.beginPath();
 
@@ -841,6 +966,8 @@ import { bind, unbind, trigger, css } from './helpers.js';
                     ctx.moveTo(startx + PANHINT_LENGTH_CONSTANT, starty);
                     ctx.lineTo(startx - PANHINT_LENGTH_CONSTANT, starty);
                 } else {
+                    var endx = Math.round(panAxes && panAxes[0].direction === 'y' ? panHint.start.x : panHint.end.x),
+                        endy = Math.round(panAxes && panAxes[0].direction === 'x' ? panHint.start.y : panHint.end.y);
                     var dirX = starty === endy;
 
                     ctx.moveTo(startx - (dirX ? 0 : PANHINT_LENGTH_CONSTANT), starty - (dirX ? PANHINT_LENGTH_CONSTANT : 0));
@@ -857,6 +984,7 @@ import { bind, unbind, trigger, css } from './helpers.js';
             }
         }
 
+        /** @param {number} touchPointX @param {number} touchPointY */
         plot.getTouchedAxis = function(touchPointX, touchPointY) {
             var placeholderRect = plot.getPlaceholder().getBoundingClientRect();
             var ec = { left: placeholderRect.left + window.scrollX, top: placeholderRect.top + window.scrollY };